requirements.txt
aioboto3==14.3.0
aiobotocore==2.22.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aioitertools==0.12.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
boto3==1.37.3
botocore==1.37.3
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
idna==3.10
iso8601==2.1.0
jmespath==1.0.1
multidict==6.4.4
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
pypika-tortoise==0.6.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-multipart==0.0.20
pytz==2025.2
redis==6.2.0
s3transfer==0.11.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tortoise-orm==0.25.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.3
wrapt==1.17.2
yarl==1.20.0
